edit waiter page
