+++ order-system-ionic/src/pages/Waiter/Waiter.module.scss
@@ -1,0 +1,19 @@
+.centerContainer {
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-evenly;
+    height: 100%;
+  }
+  
+.centerText {
+font-size: 1.5rem;
+margin: 0px 30px;
+font-weight: bold;
+}
+
+.resetButton {
+margin-top: 20px;
+}
+
